guardrails/input_guardrail/pyproject.toml
[project]
name = "input-guardrail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
